back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "click==8.1.8",
    "fastapi==0.115.12",
    "h11==0.14.0",
    "idna==3.10",
    "moviepy>=2.1.2",
    "mpmath==1.3.0",
    "numpy>=2.2.4",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusparselt-cu12==0.6.2",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "opencv-python>=4.11.0.86",
    "packaging==24.2",
    "pandas>=2.2.3",
    "pillow==10.4.0",
    "platformdirs==4.3.7",
    "psutil==7.0.0",
    "pycparser==2.22",
    "pydantic==2.11.1",
    "pydantic-core==2.33.0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests>=2.32.3",
    "safetensors==0.5.3",
    "setuptools==78.1.0",
    "sniffio==1.3.1",
    "starlette==0.46.1",
    "sympy==1.13.1",
    "threadpoolctl==3.6.0",
    "torch>=2.6.0",
    "tqdm==4.67.1",
    "triton==3.2.0",
    "typing-extensions==4.13.0",
    "typing-inspection==0.4.0",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
]
